requirements.txt
beautifulsoup4==4.10.0
category-encoders==2.3.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
Flask==2.0.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
lxml==4.6.4
MarkupSafe==2.0.1
numpy==1.21.4
pandas==1.3.4
patsy==0.5.2
psycopg2==2.9.2
pymongo==4.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
soupsieve==2.3.1
statsmodels==0.13.1
threadpoolctl==3.0.0
urllib3==1.26.7
Werkzeug==2.0.2
wincertstore==0.2
